refactor(api): move v1 handlers
