+++ tests/test_api.py
@@ -7,7 +7,7 @@ import pytest
 from fastapi import HTTPException
 from starlette.requests import Request
 from sqlalchemy import text
-from app.main import compute_signature, verify_hmac
+from app.dependencies import compute_signature, verify_hmac
 from app.services.protocols import import_csv_to_db
 
 
@@ -18,7 +18,7 @@ def stub_upload(monkeypatch):
         return "1/stub.jpg"
 
     monkeypatch.setattr("app.services.storage.upload_photo", _stub)
-    monkeypatch.setattr("app.main.upload_photo", _stub)
+    monkeypatch.setattr("app.controllers.v1.upload_photo", _stub)
     # reset usage to avoid 402 errors between tests
     from app.db import SessionLocal
     from app.models import PhotoUsage
@@ -200,7 +200,7 @@ def test_diagnose_gpt_timeout(monkeypatch, client):
     def _fail(_key: str):
         raise TimeoutError("timeout")
 
-    monkeypatch.setattr("app.main.call_gpt_vision_stub", _fail)
+    monkeypatch.setattr("app.controllers.v1.call_gpt_vision_stub", _fail)
 
     resp = client.post(
         "/v1/ai/diagnose",
@@ -747,7 +747,7 @@ def test_diagnose_json_no_protocol_beta(client):
 
 
 def test_free_monthly_limit_env(monkeypatch, client):
-    monkeypatch.setattr("app.main.FREE_MONTHLY_LIMIT", 1)
+    monkeypatch.setattr("app.controllers.v1.FREE_MONTHLY_LIMIT", 1)
     resp = client.get("/v1/limits", headers=HEADERS)
     assert resp.status_code == 200
     assert resp.json()["limit_monthly_free"] == 1
@@ -790,8 +790,8 @@ def test_sixth_diagnose_call_returns_402(client):
 
 
 def test_paywall_disabled_returns_200(monkeypatch, client):
-    monkeypatch.setattr("app.main.FREE_MONTHLY_LIMIT", 1)
-    monkeypatch.setattr("app.main.PAYWALL_ENABLED", False)
+    monkeypatch.setattr("app.controllers.v1.FREE_MONTHLY_LIMIT", 1)
+    monkeypatch.setattr("app.controllers.v1.PAYWALL_ENABLED", False)
 
     first = client.post(
         "/v1/ai/diagnose",
@@ -822,7 +822,7 @@ def test_diagnose_old_sqlite_fallback(monkeypatch, client):
 
 
 def test_pro_expired_event_logged(monkeypatch, client):
-    monkeypatch.setattr("app.main.FREE_MONTHLY_LIMIT", 0)
+    monkeypatch.setattr("app.controllers.v1.FREE_MONTHLY_LIMIT", 0)
     from app.db import SessionLocal
     from app.models import Event
     past = datetime(2024, 1, 1, tzinfo=timezone.utc)
